change the flow of discount calcualtions
